backend_django/pyproject.toml
[tool.poetry]
name = "backend-django"
version = "0.1.0"
description = ""
authors = ["dapt <diegodapt18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
pynvim = "^0.4.3"
django-cors-headers = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
